requirements.txt
streamlit
pydub
moviepy
deepgram-sdk
requests
python-dotenv
soundfile
